third_party/WebKit/LayoutTests/inspector-protocol/accessibility/accessibility-ignoredNodes.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/dom-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="accessibility-dumpAccessibilityNodes.js"></script>
<script>

function test()
{
    InspectorTest.sendCommand("DOM.getDocument", {}, InspectorTest.dumpAccessibilityNodesBySelectorAndCompleteTest.bind(null, '[data-dump]'));
}
</script>
</head>
<body onLoad="runTest();">
    <div data-dump>Non-hidden div for comparison</div>
    <div role="tree">
        <div data-dump role="treeitem">non-hidden treeitem</div>
        <div data-dump >hidden non-treeitem</div>
    </div>
    <div role="img">
        <svg data-dump>
            <!-- Children of img role are presentational -->
            <circle xmlns:svg="http://www.w3.org/2000/svg" cx="150px" cy="100px" r="50px" fill="#ff0000" stroke="#000000" stroke-width="5px"/>
        </svg>
    </div>
    <button>
        <span data-dump>Buttons are leaf nodes</span>
    </button>
    <div data-dump aria-hidden="true">
        <div data-dump>Descendant of aria-hidden node</div>
    </div>
    <img data-dump alt="" src="foo.png">
    <ol role="none" data-dump><!-- list is presentational -->
      <li data-dump>List item also presentational</li>
      <div data-dump>Div in list isn't presentational</div>
    </ol>

    <label for="checkbox" data-dump><span data-dump>Content within label refers to label container</span></label>
    <input type="checkbox" id="checkbox">

    <div style="display: none" data-dump>Non-rendered div</div>

    <canvas style="height: 1px; width: 1px;q" data-dump></canvas>
</body>
</html>
